// lib/types/components/textarea.d.ts
import { type ReactNode } from "react";
import { type TColors } from "../generic";

export type TTextareaVariant = "filled" | "outlined";
export type TTextareaSize = "sm" | "md" | "lg";
export type TTextareaColors = TColors;
export type TTextareaFullWidth = boolean;
export type TTextareaDisabled = boolean;
export type TTextareaError = boolean;
export type TTextareaSuccess = boolean;
export type TTextareaLabel = string;
export type TTextareaHelperText = string;
export type TTextareaErrorText = string;
export type TTextareaSuccessText = string;
export type TTextareaStartIcon = ReactNode;
export type TTextareaEndIcon = ReactNode;
export type TTextareaClassname = string;